fix(windows): close residual Unicode and LLP64 gaps

+++ src/t_stream.c
@@ -6332,7 +6332,7 @@ static int createIdempotencyHash(robj **argv, int64_t numfields, XXH128_hash_t *
         if (err != XXH_OK) goto cleanup;
 
         /* Hash the field */
-        long field_len;
+        size_t field_len;
         unsigned char *field_data = getObjectReadOnlyString(field, &field_len, llbuf);
         err = XXH3_128bits_update(state, field_data, field_len);
         if (err != XXH_OK) goto cleanup;
@@ -6342,7 +6342,7 @@ static int createIdempotencyHash(robj **argv, int64_t numfields, XXH128_hash_t *
         if (err != XXH_OK) goto cleanup;
 
         /* Hash the value */
-        long value_len;
+        size_t value_len;
         unsigned char *value_data = getObjectReadOnlyString(value, &value_len, llbuf);
         err = XXH3_128bits_update(state, value_data, value_len);
         if (err != XXH_OK) goto cleanup;
